+++ crow/__init__.py
@@ -1,1 +1,4 @@
+from . import uitabs
+from . import utils
+
 __version__ = "2.0.0"
